// src/components/types/infoAuthTypes.ts
export type AuthoInfo = {
  id: number;
  email: string;
  token: string;
};
export interface ISignIn {
  email: string;
  password: string;
}
export interface ISignUp {
  email: string;
  password: string;
  confirmPassword: string;
}

export interface IError {
  message: string;
  code: number;
}